Updated to UVVM v3 BETA 2024.09.10

+++ bitvis_uart/script/maintenance_script/test.py
@@ -3,6 +3,7 @@ import os
 import shutil
 from itertools import product
 
+
 try:
     from hdlregression import HDLRegression
 except:
@@ -15,40 +16,39 @@ def cleanup(msg='Cleaning up...'):
 
     sim_path = os.getcwd()
 
-    for files in os.listdir(sim_path):
-        path = os.path.join(sim_path, files)
-        try:
-            shutil.rmtree(path)
-        except:
-            os.remove(path)
+    # Check if the current directory is 'sim'
+    if os.path.basename(sim_path) == 'sim':
+        for files in os.listdir(sim_path):
+            path = os.path.join(sim_path, files)
+            try:
+                shutil.rmtree(path)
+            except:
+                os.remove(path)
+    else:
+        print('Current directory is not "sim". Skipping cleanup.')
 
 
 print('Verify Bitvis UART DUT')
 
 cleanup('Removing any previous runs.')
 
-hr = HDLRegression(simulator='modelsim')
+hr = HDLRegression()
 
 # Add util, fw and VIP Scoreboard
 hr.add_files("../../../uvvm_util/src/*.vhd", "uvvm_util")
 hr.add_files("../../../uvvm_vvc_framework/src/*.vhd", "uvvm_vvc_framework")
-hr.add_files("../../../bitvis_vip_scoreboard/src/*.vhd",
-             "bitvis_vip_scoreboard")
+hr.add_files("../../../bitvis_vip_scoreboard/src/*.vhd", "bitvis_vip_scoreboard")
 
 # Add other VIPs in the TB
 #  - SBI VIP
 hr.add_files("../../../bitvis_vip_sbi/src/*.vhd", "bitvis_vip_sbi")
-hr.add_files(
-    "../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_sbi")
+hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_sbi")
 #  - UART VIP
 hr.add_files("../../../bitvis_vip_uart/src/*.vhd", "bitvis_vip_uart")
-hr.add_files(
-    "../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_uart")
+hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_uart")
 #  - Clock Generator VVC
-hr.add_files("../../../bitvis_vip_clock_generator/src/*.vhd",
-             "bitvis_vip_clock_generator")
-hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd",
-             "bitvis_vip_clock_generator")
+hr.add_files("../../../bitvis_vip_clock_generator/src/*.vhd", "bitvis_vip_clock_generator")
+hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_vip_clock_generator")
 # Add DUT
 hr.add_files("../../../bitvis_uart/src/*.vhd", "bitvis_uart")
 
@@ -62,7 +62,7 @@ simulator_name = hr.settings.get_simulator_name()
 if simulator_name in ['MODELSIM', 'RIVIERA']:
     sim_options = '-t ns'
     # Set compile options
-    default_options = ["-suppress", "1346,1246,1236,1090", "-2008"]
+    default_options = ["-suppress", "1346,1246,1236", "-2008"]
     hr.set_simulator(simulator=simulator_name, com_options=default_options)
 
 hr.start(sim_options=sim_options)
