+++ bitvis_vip_axistream/script/maintenance_script/test.py
@@ -16,31 +16,22 @@ def cleanup(msg='Cleaning up...'):
 
     sim_path = os.getcwd()
 
-    for files in os.listdir(sim_path):
-        path = os.path.join(sim_path, files)
-        try:
-            shutil.rmtree(path)
-        except:
-            os.remove(path)
+    # Check if the current directory is 'sim'
+    if os.path.basename(sim_path) == 'sim':
+        for files in os.listdir(sim_path):
+            path = os.path.join(sim_path, files)
+            try:
+                shutil.rmtree(path)
+            except:
+                os.remove(path)
+    else:
+        print('Current directory is not "sim". Skipping cleanup.')
+
 
 def create_config(data_widths, user_widths, id_widths, dest_widths, include_tuser=False, use_setup_and_hold=False):
     config = []
-
-    # Need to convert boolean to int - and handle lists
-    if isinstance(include_tuser, list):
-        for idx, item in enumerate(include_tuser):
-            include_tuser[idx] = 1 if item is True else 0
-    else:
-        include_tuser = 1 if include_tuser is True else 0
-    # Need to convert boolean to int - and handle lists
-    if isinstance(use_setup_and_hold, list):
-        for idx, item in enumerate(use_setup_and_hold):
-            use_setup_and_hold[idx] = 1 if item is True else 0
-    else:
-        use_setup_and_hold = 1 if use_setup_and_hold is True else 0
-    
     for data_width, user_width, id_width, dest_width, tuser, setup_and_hold in product(data_widths, user_widths, id_widths, dest_widths, include_tuser, use_setup_and_hold):
-      config.append([str(data_width), str(user_width), str(id_width), str(dest_width), str(tuser), str(setup_and_hold)])
+        config.append([str(data_width), str(user_width), str(id_width), str(dest_width), str(tuser), str(setup_and_hold)])
     return config
 
 
@@ -48,7 +39,7 @@ print('Verify Bitvis VIP AXI Stream')
 
 cleanup('Removing any previous runs.')
 
-hr = HDLRegression(simulator='modelsim')
+hr = HDLRegression()
 
 # Add util, fw and VIP Scoreboard
 hr.add_files("../../../uvvm_util/src/*.vhd", "uvvm_util")
@@ -59,28 +50,28 @@ hr.add_files("../../../bitvis_vip_scoreboard/src/*.vhd", "bitvis_vip_scoreboard"
 configs = create_config(data_widths=[32], user_widths=[8], id_widths=[7], dest_widths=[4], include_tuser=[False], use_setup_and_hold=[True, False])
 for config in configs:
     hr.add_generics(entity="axistream_bfm_tb",
-                         generics=["GC_DATA_WIDTH", config[0],"GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
+                    generics=["GC_DATA_WIDTH", config[0], "GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
 
-configs = create_config(data_widths=[8, 16, 24, 64, 128], user_widths=[8], id_widths=[8], dest_widths=[4], include_tuser=[True], use_setup_and_hold=[True])
+configs = create_config(data_widths=[8, 16, 24, 64, 128], user_widths=[8, 16], id_widths=[8], dest_widths=[4], include_tuser=[True], use_setup_and_hold=[True])
 for config in configs:
     hr.add_generics(entity="axistream_vvc_tb",
-                         generics=["GC_DATA_WIDTH", config[0],"GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
+                    generics=["GC_DATA_WIDTH", config[0], "GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
 
 configs = create_config(data_widths=[32], user_widths=[1, 5], id_widths=[3, 7], dest_widths=[1], include_tuser=[True], use_setup_and_hold=[False])
 for config in configs:
     hr.add_generics(entity="axistream_vvc_tb",
-                         generics=["GC_DATA_WIDTH", config[0],"GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
+                    generics=["GC_DATA_WIDTH", config[0], "GC_USER_WIDTH", config[1], "GC_ID_WIDTH", config[2], "GC_DEST_WIDTH", config[3], "GC_INCLUDE_TUSER", config[4], "GC_USE_SETUP_AND_HOLD", config[5]])
 
 hr.add_generics(entity="axistream_vvc_tb",
-                     generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 1, "GC_ID_WIDTH", 1, "GC_DEST_WIDTH", 1, "GC_INCLUDE_TUSER", False, "GC_USE_SETUP_AND_HOLD", True])
+                generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 1, "GC_ID_WIDTH", 1, "GC_DEST_WIDTH", 1, "GC_INCLUDE_TUSER", False, "GC_USE_SETUP_AND_HOLD", True])
 
-hr.add_generics(entity="axistream_multiple_vvc_tb",
-                     generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 1, "GC_ID_WIDTH", 1, "GC_DEST_WIDTH", 1, "GC_INCLUDE_TUSER", True, "GC_USE_SETUP_AND_HOLD", True])
+hr.add_generics(entity="axistream_vvc_multiple_tb",
+                generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 1, "GC_ID_WIDTH", 1, "GC_DEST_WIDTH", 1, "GC_INCLUDE_TUSER", True, "GC_USE_SETUP_AND_HOLD", True])
 
-hr.add_generics(entity="axistream_multiple_vvc_tb",
-                     generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 8, "GC_ID_WIDTH", 7, "GC_DEST_WIDTH", 4, "GC_INCLUDE_TUSER", True, "GC_USE_SETUP_AND_HOLD", False])
+hr.add_generics(entity="axistream_vvc_multiple_tb",
+                generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 8, "GC_ID_WIDTH", 7, "GC_DEST_WIDTH", 4, "GC_INCLUDE_TUSER", True, "GC_USE_SETUP_AND_HOLD", False])
 
-hr.add_generics(entity="axistream_width_vvc_tb",
+hr.add_generics(entity="axistream_vvc_width_tb",
                 generics=["GC_DATA_WIDTH", 32, "GC_USER_WIDTH", 1, "GC_ID_WIDTH", 1, "GC_DEST_WIDTH", 1, "GC_INCLUDE_TUSER", True, "GC_USE_SETUP_AND_HOLD", True])
 
 # Add src files
@@ -91,13 +82,14 @@ hr.add_files("../../../uvvm_vvc_framework/src_target_dependent/*.vhd", "bitvis_v
 hr.add_files("../../tb/maintenance_tb/*.vhd", "bitvis_vip_axistream")
 
 sim_options = None
-default_options = []
 simulator_name = hr.settings.get_simulator_name()
-if simulator_name in ['MODELSIM', 'RIVIERA']:
-    sim_options = '-t ns'
-    # Set compile options
-    default_options = ["-suppress", "1346,1246,1236,1090", "-2008"]
-    hr.set_simulator(simulator=simulator_name, com_options=default_options)
+# Set simulator name and compile options
+if simulator_name in ["MODELSIM", "RIVIERA"]:
+    sim_options = "-t ns"
+    com_options = ["-suppress", "1346,1246,1236", "-2008"]
+    hr.set_simulator(simulator=simulator_name, com_options=com_options)
+elif simulator_name == "NVC":
+    sim_options = ["-M64m", "-H2g"]
 
 hr.start(sim_options=sim_options)
